+++ packages/tv-ad-admin-next/app/dashboard/page.tsx
@@ -2,12 +2,10 @@
 
 // import { useState, useEffect } from 'react'
 
-import Link from 'next/link'
 
 import Image from 'next/image'
+import Link from 'next/link'
 
-import fileIcon from '@/assets/icons/file.svg'
-import uploadIcon from '@/assets/icons/upload.svg'
 import StatusCard from '@/components/dashboard/status-card'
 import PageHeader from '@/components/shared/page-header'
 import PageMain from '@/components/shared/page-main'
@@ -19,6 +17,8 @@ import {
   CardContent,
 } from '@/components/ui/card'
 import { OrderStatusMap } from '@/constants'
+import fileIcon from '@/public/icons/file.svg'
+import uploadIcon from '@/public/icons/upload.svg'
 // import { mockOrderData } from '@/mocks/mockData'
 
 // type StatusStats = { status: string; count: number }
