+++ packages/tv-ad-admin-next/components/order/order-actions.tsx
@@ -2,12 +2,12 @@ import { useState } from 'react'
 
 import Link from 'next/link'
 
-import DoneCircleIcon from '@/assets/icons/done-circle.svg'
-import EditIcon from '@/assets/icons/edit.svg'
-import UploadIcon from '@/assets/icons/upload.svg'
 import { Button } from '@/components/ui/button'
 import { ORDER_STATUS } from '@/constants'
 import { type OrderRecord } from '@/mocks/mockData'
+import DoneCircleIcon from '@/public/icons/done-circle.svg'
+import EditIcon from '@/public/icons/edit.svg'
+import UploadIcon from '@/public/icons/upload.svg'
 
 type OrderActionsProps = {
   order: OrderRecord
@@ -200,10 +200,10 @@ export function OrderActions({ order, className = '' }: OrderActionsProps) {
               actionContent.buttonText === '上傳素材'
                 ? handleUploadClick
                 : actionContent.buttonText === '確認'
-                ? handleConfirmClick
-                : actionContent.buttonText === '設定排播日期'
-                ? handleSettingScheduleClick
-                : undefined
+                  ? handleConfirmClick
+                  : actionContent.buttonText === '設定排播日期'
+                    ? handleSettingScheduleClick
+                    : undefined
             }
             disabled={isUploading || isConfirming}
           >
@@ -211,8 +211,8 @@ export function OrderActions({ order, className = '' }: OrderActionsProps) {
             {isUploading && actionContent.buttonText === '上傳素材'
               ? '上傳中...'
               : isConfirming && actionContent.buttonText === '確認'
-              ? '確認中...'
-              : actionContent.buttonText}
+                ? '確認中...'
+                : actionContent.buttonText}
           </Button>
         )}
         {actionContent.secondaryButton && (
